requirements.txt
pytest == 7.4.3
openpyxl == 3.1.2
PyPDF2 == 3.0.1